+++ singularity-eos/eos/eos_stiff.hpp
@@ -166,7 +166,6 @@ class StiffGas : public EosBase<StiffGas> {
   // overloads that use the scalar versions declared here
   SG_ADD_BASE_CLASS_USINGS(StiffGas)
   SG_ADD_DEFAULT_MEAN_ATOMIC_FUNCTIONS(_AZbar)
-  PORTABLE_INLINE_FUNCTION
 
   static constexpr unsigned long PreferredInput() { return _preferred_input; }
   PORTABLE_INLINE_FUNCTION void PrintParams() const {